Moved logging from the endpoints to the model layer
