+++ ghost/core/core/server/api/endpoints/automated-emails.js
@@ -1,45 +1,10 @@
 const tpl = require('@tryghost/tpl');
 const errors = require('@tryghost/errors');
-const logging = require('@tryghost/logging');
 const models = require('../../models');
 const memberWelcomeEmailService = require('../../services/member-welcome-emails/service');
-const {MEMBER_WELCOME_EMAIL_SLUGS} = require('../../services/member-welcome-emails/constants');
-
-const MEMBER_WELCOME_EMAIL_SLUG_SET = new Set(Object.values(MEMBER_WELCOME_EMAIL_SLUGS));
 
 const messages = {
     automatedEmailNotFound: 'Automated email not found.'
-};
-
-const logWelcomeEmailStatusChange = (model) => {
-    if (!model?.id) {
-        return;
-    }
-
-    const slug = model.get('slug');
-
-    if (!MEMBER_WELCOME_EMAIL_SLUG_SET.has(slug)) {
-        return;
-    }
-
-    const previousStatus = model.previous('status');
-    const currentStatus = model.get('status');
-    const isNewModel = previousStatus === undefined;
-    const isEnableTransition = currentStatus === 'active' && (isNewModel || previousStatus === 'inactive');
-    const isDisableTransition = previousStatus === 'active' && currentStatus === 'inactive';
-
-    if (!isEnableTransition && !isDisableTransition) {
-        return;
-    }
-
-    logging.info({
-        system: {
-            event: isEnableTransition ? 'welcome_email.enabled' : 'welcome_email.disabled',
-            automated_email_id: model.id,
-            slug,
-            enabled: currentStatus === 'active'
-        }
-    }, isEnableTransition ? 'Welcome email enabled' : 'Welcome email disabled');
 };
 
 /** @type {import('@tryghost/api-framework').Controller} */
@@ -95,9 +60,7 @@ const controller = {
         permissions: true,
         async query(frame) {
             const data = frame.data.automated_emails[0];
-            const model = await models.AutomatedEmail.add(data, frame.options);
-            logWelcomeEmailStatusChange(model);
-            return model;
+            return models.AutomatedEmail.add(data, frame.options);
         }
     },
 
@@ -124,8 +87,6 @@ const controller = {
                     message: tpl(messages.automatedEmailNotFound)
                 });
             }
-
-            logWelcomeEmailStatusChange(model);
 
             return model;
         }
