+++ ghost/core/core/server/models/automated-email.js
@@ -1,6 +1,10 @@
 const ghostBookshelf = require('./base');
+const logging = require('@tryghost/logging');
 const urlUtils = require('../../shared/url-utils');
 const lexicalLib = require('../lib/lexical');
+const {MEMBER_WELCOME_EMAIL_SLUGS} = require('../services/member-welcome-emails/constants');
+
+const MEMBER_WELCOME_EMAIL_SLUG_SET = new Set(Object.values(MEMBER_WELCOME_EMAIL_SLUGS));
 
 const AutomatedEmail = ghostBookshelf.Model.extend({
     tableName: 'automated_emails',
@@ -33,6 +37,37 @@ const AutomatedEmail = ghostBookshelf.Model.extend({
         }
 
         return attrs;
+    },
+
+    onSaved(model) {
+        if (!model?.id) {
+            return;
+        }
+
+        const slug = model.get('slug');
+
+        if (!MEMBER_WELCOME_EMAIL_SLUG_SET.has(slug)) {
+            return;
+        }
+
+        const previousStatus = model.previous('status');
+        const currentStatus = model.get('status');
+        const isNewModel = previousStatus === undefined;
+        const isEnableTransition = currentStatus === 'active' && (isNewModel || previousStatus === 'inactive');
+        const isDisableTransition = previousStatus === 'active' && currentStatus === 'inactive';
+
+        if (!isEnableTransition && !isDisableTransition) {
+            return;
+        }
+
+        logging.info({
+            system: {
+                event: isEnableTransition ? 'welcome_email.enabled' : 'welcome_email.disabled',
+                automated_email_id: model.id,
+                slug,
+                enabled: currentStatus === 'active'
+            }
+        }, isEnableTransition ? 'Welcome email enabled' : 'Welcome email disabled');
     }
 });
